requirements.txt
torch
torchvision
numpy
pandas
matplotlib
opencv-python
Pillow
tqdm
scikit-learn
tensorboard
PyYAML
seaborn
pyquaternion
tqdm
wandb
ultralytics
python-magic
python-magic-bin
